Dev2

+++ ThMoCo.WebApp/Models/AppUserDTO.cs
@@ -7,7 +7,7 @@ public class AppUserDTO
     public string? Email { get; set; }
     public string? PhotoUrl { get; set; }
     public string? PhoneNumber { get; set; }
-    public double? Fund { get; set; }
+    public decimal? Fund { get; set; }
     public DateTime? UpdatedAt { get; set; }
     public DateTime? LastLogin { get; set; }
     public string? Provider { get; set; } // e.g., "auth0", "google", "facebook"
